+++ src/cliques/cliquer_wrapper.c
@@ -231,15 +231,16 @@ igraph_error_t igraph_i_cliquer_histogram(const igraph_t *graph, igraph_vector_t
 /* Call function for each clique. */
 
 struct callback_data {
+    igraph_vector_int_t *clique;
     igraph_clique_handler_t *handler;
     void *arg;
 };
 
 static igraph_error_t callback_callback(set_t s, graph_t *g, clique_options *opt) {
-    igraph_vector_int_t *clique;
     struct callback_data *cd;
     int i;
     igraph_integer_t j;
+    igraph_error_t retval;
 
     IGRAPH_UNUSED(g);
 
@@ -247,24 +248,25 @@ static igraph_error_t callback_callback(set_t s, graph_t *g, clique_options *opt
 
     cd = (struct callback_data *) opt->user_data;
 
-    clique = (igraph_vector_int_t *) malloc(sizeof(igraph_vector_int_t));
-    if (clique == 0) {
-        IGRAPH_ERROR("storing cliques failed", IGRAPH_ENOMEM); /* LCOV_EXCL_LINE */
-    }
-    IGRAPH_CHECK(igraph_vector_int_init(clique, set_size(s)));
+    IGRAPH_CHECK(igraph_vector_int_resize(cd->clique, set_size(s)));
 
     i = -1; j = 0;
     while ((i = set_return_next(s, i)) >= 0) {
-        VECTOR(*clique)[j++] = i;
+        VECTOR(*cd->clique)[j++] = i;
     }
 
-    return (*(cd->handler))(clique, cd->arg);
+    retval = (*(cd->handler))(cd->clique, cd->arg);
+
+    igraph_vector_int_destroy(cd->clique);
+
+    return retval;
 }
 
 igraph_error_t igraph_i_cliquer_callback(const igraph_t *graph,
                               igraph_integer_t min_size, igraph_integer_t max_size,
                               igraph_clique_handler_t *cliquehandler_fn, void *arg) {
     graph_t *g;
+    igraph_vector_int_t current_clique;
     struct callback_data cd;
     igraph_integer_t vcount = igraph_vcount(graph);
 
@@ -290,6 +292,9 @@ igraph_error_t igraph_i_cliquer_callback(const igraph_t *graph,
     IGRAPH_CHECK(igraph_to_cliquer(graph, &g));
     IGRAPH_FINALLY(graph_free, g);
 
+    IGRAPH_VECTOR_INT_INIT_FINALLY(&current_clique, min_size);
+
+    cd.clique = &current_clique;
     cd.handler = cliquehandler_fn;
     cd.arg = arg;
     igraph_cliquer_opt.user_data = &cd;
@@ -297,8 +302,9 @@ igraph_error_t igraph_i_cliquer_callback(const igraph_t *graph,
 
     IGRAPH_CHECK(clique_unweighted_find_all(g, (int) min_size, (int) max_size, /* maximal= */ FALSE, &igraph_cliquer_opt, NULL));
 
+    igraph_vector_int_destroy(&current_clique);
     graph_free(g);
-    IGRAPH_FINALLY_CLEAN(1);
+    IGRAPH_FINALLY_CLEAN(2);
 
     return IGRAPH_SUCCESS;
 }